二重送信チェックのサンプルコードを見直し
