+++ samples/web/double-submission/src/main/java/keel/doublesubmission/UserForm.java
@@ -1,0 +1,31 @@
+package keel.doublesubmission;
+
+import javax.validation.constraints.Max;
+import javax.validation.constraints.Min;
+import javax.validation.constraints.NotEmpty;
+
+public class UserForm {
+
+    @NotEmpty
+    private String name;
+
+    @Min(1)
+    @Max(150)
+    private Integer age;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public Integer getAge() {
+        return age;
+    }
+
+    public void setAge(Integer age) {
+        this.age = age;
+    }
+}
