+++ samples/web/double-submission/src/test/java/keel/doublesubmission/UserControllerTest.java
@@ -9,7 +9,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.mock.web.MockHttpSession;
 import org.springframework.test.web.servlet.MockMvc;
@@ -21,16 +20,11 @@ public class UserControllerTest {
     @Autowired
     private MockMvc mockMvc;
 
-    @MockBean
-    private UserService service;
-
     @Test
     public void トークンが生成されているかを確認するテスト() throws Exception {
-        mockMvc
-                .perform(post("/user/confirm")
+        mockMvc.perform(post("/user/confirm")
                         .param("name", "taro")
-                        .param("age", "25")
-                )
+                        .param("age", "25"))
                 .andExpect(status().isOk())
                 .andExpect(model().hasNoErrors())
                 .andExpect(view().name("user/confirm"))
@@ -40,11 +34,10 @@ public class UserControllerTest {
 
     @Test
     public void トークンをリクエスト送信しない場合のテスト() throws Exception {
-        mockMvc
-                .perform(post("/user/create")
+        mockMvc.perform(
+                post("/user/create")
                         .param("name", "taro")
-                        .param("age", "25")
-                )
+                        .param("age", "25"))
                 .andExpect(status().isBadRequest())
                 .andExpect(view().name("error/token-error"));
     }
@@ -52,11 +45,10 @@ public class UserControllerTest {
     @Test
     public void トークンをリクエスト送信する場合のテスト() throws Exception {
 
-        final MockHttpServletRequest request = mockMvc
-                .perform(post("/user/confirm")
+        MockHttpServletRequest request = mockMvc.perform(
+                post("/user/confirm")
                         .param("name", "taro")
-                        .param("age", "25")
-                )
+                        .param("age", "25"))
                 .andExpect(status().isOk())
                 .andReturn()
                 .getRequest();
@@ -64,13 +56,44 @@ public class UserControllerTest {
         MockHttpSession session = (MockHttpSession) request.getSession();
         TransactionToken transactionToken = (TransactionToken) request.getAttribute(TransactionTokenInterceptor.NEXT_TOKEN_REQUEST_ATTRIBUTE_NAME);
 
-        mockMvc
-                .perform(
-                        post("/user/create")
-                                .param("name", "taro")
-                                .param("age", "25")
-                                .param("_TRANSACTION_TOKEN", transactionToken.getTokenString())
-                                .session(session))
+        mockMvc.perform(
+                post("/user/create")
+                        .param("name", "taro")
+                        .param("age", "25")
+                        .param("_TRANSACTION_TOKEN", transactionToken.getTokenString())
+                        .session(session))
                 .andExpect(redirectedUrl("/user/complete"));
+    }
+
+    @Test
+    public void トークンが再利用されるとトークンチェックでエラーになる() throws Exception {
+
+        final MockHttpServletRequest request = mockMvc.perform(
+                post("/user/confirm")
+                    .param("name", "error")
+                    .param("age", "25"))
+                .andExpect(status().isOk())
+                .andReturn()
+                .getRequest();
+
+        MockHttpSession session = (MockHttpSession) request.getSession();
+        TransactionToken transactionToken = (TransactionToken) request.getAttribute(TransactionTokenInterceptor.NEXT_TOKEN_REQUEST_ATTRIBUTE_NAME);
+
+        mockMvc.perform(
+                post("/user/create")
+                        .param("name", "taro")
+                        .param("age", "25")
+                        .param("_TRANSACTION_TOKEN", transactionToken.getTokenString())
+                        .session(session))
+                .andExpect(model().hasNoErrors());
+
+        mockMvc.perform(
+                post("/user/create")
+                        .param("name", "taro")
+                        .param("age", "26")
+                        .param("_TRANSACTION_TOKEN", transactionToken.getTokenString())
+                        .session(session))
+                .andExpect(status().isBadRequest())
+                .andExpect(view().name("error/token-error"));
     }
 }
